third_party/blink/web_tests/platform/generic/editing/deleting/delete-svg-001-expected.txt

PASS sample.textContent.charCodeAt(0) is 160
PASS successfullyParsed is true

TEST COMPLETE
